@@ -1,11 +1,12 @@
 module github.com/binaryholdings/cosmos-pruner
 
-go 1.17
+go 1.22
+
+toolchain go1.24.1
 
 require (
 	github.com/cosmos/cosmos-sdk v0.45.2-0.20220209090836-4161d2b6022a
 	github.com/cosmos/iavl v0.17.3
-	github.com/cosmos/ibc-go/v2 v2.0.2
 	github.com/gogo/protobuf v1.3.3
 	github.com/neilotoole/errgroup v0.1.5
 	github.com/pkg/errors v0.9.1
